add lessons

+++ controllers/TicketController.js
@@ -1,5 +1,6 @@
 import StudentModel from '../models/Student.js';
 import Ticketmodel from '../models/Ticket.js';
+import LessonModel from '../models/Lesson.js';
 
 export const createTicket = async (request, response) => {
     try {
@@ -71,6 +72,12 @@ export const deleteTicket = async (request, response) => {
 
         if (!ticket) {
             return response.status(400).json('Ticket not found');
+        }
+
+        const lessons = await LessonModel.find({ticket: ticket._id});
+
+        if (lessons.length > 0) {
+            return response.status(200).json({message: 'Ticket has lessons'})
         }
 
         const deletad = await Ticketmodel.deleteOne({_id: id});
